requirements.txt
redis
aiohttp
aiogram
pytz
python-dotenv